backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
djongo
django-cors-headers
gunicorn
Django>=4.0
djangorestframework
djongo
pymongo
djangorestframework-simplejwt
django-cors-headers

